created addReviewReducer module
